refactor: move to modular.ts
